# documentation/openapi.json
{
  "components" : {
    "schemas" : {
      "PageRestElementDeFabrication" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/RestElementDeFabrication"
            },
            "type" : "array"
          },
          "currentPage" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElementsCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "content", "currentPage", "pageSize", "totalElementsCount" ],
        "type" : "object"
      },
...
      "PageRestPosteDeTravail" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/RestPosteDeTravail"
            },
            "type" : "array"
          },
          "currentPage" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElementsCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "content", "currentPage", "pageSize", "totalElementsCount" ],
        "type" : "object"
      },
...
      "RestCreationElementDeFabrication" : {
        "properties" : {
          "description" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "reference" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "type" : {
            "enum" : [ "ORDRE_DE_FABRICATION", "PRODUIT" ],
            "type" : "string"
          }
        },
        "required" : [ "type" ],
        "type" : "object"
      },
...
      "RestCreationPosteDeTravail" : {
        "description" : "Declaration d'un poste de travail.",
        "properties" : {
          "coutHoraire" : {
            "description" : "Cout horaire du poste, destine au cout de revient. Facultatif : toutes les entreprises ne le valorisent pas.",
            "example" : 45.5,
            "minimum" : 0,
            "type" : "number"
          },
          "libelle" : {
            "description" : "Nom du poste, unique dans l'entreprise.",
            "example" : "Tour 1",
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "nature" : {
            "description" : "Metier qui s'exerce sur ce poste. Obligatoire : un poste est declare pour dire quel travail s'y fait.",
            "example" : "tournage",
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "libelle", "nature" ],
        "type" : "object"
      },
      "RestElementDeFabrication" : {
        "properties" : {
          "dateDeCreation" : {
            "format" : "date-time",
            "type" : "string"
          },
          "dateDeModification" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "reference" : {
            "type" : "string"
          },
          "type" : {
            "enum" : [ "ORDRE_DE_FABRICATION", "PRODUIT" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RestModificationElementDeFabrication" : {
        "properties" : {
          "description" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "reference" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RestModificationPosteDeTravail" : {
        "description" : "Revision d'un poste de travail.",
        "properties" : {
          "coutHoraire" : {
            "description" : "Cout horaire du poste, laisse vide pour le retirer.",
            "example" : 45.5,
            "minimum" : 0,
            "type" : "number"
          },
          "libelle" : {
            "description" : "Nom du poste, unique dans l'entreprise.",
            "example" : "Tour 1",
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "nature" : {
            "description" : "Metier qui s'exerce sur ce poste.",
            "example" : "tournage",
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "libelle", "nature" ],
        "type" : "object"
      },
...
      "RestPosteDeTravail" : {
        "description" : "Un poste de travail du referentiel de l'entreprise.",
        "properties" : {
          "coutHoraire" : {
            "description" : "Cout horaire du poste, absent si l'entreprise ne le valorise pas.",
            "example" : 45.5,
            "type" : "number"
          },
          "id" : {
            "description" : "Identifiant du poste.",
            "format" : "uuid",
            "type" : "string"
          },
          "libelle" : {
            "description" : "Nom du poste tel que l'atelier le designe.",
            "example" : "Tour 1",
            "type" : "string"
          },
          "nature" : {
            "description" : "Metier qui s'exerce sur ce poste.",
            "example" : "tournage",
            "type" : "string"
          }
        },
        "required" : [ "id", "libelle", "nature" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Keycloak access token carrying the `tenant` claim.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "API of the GLM manufacturing tracking application.\n\n## Authentication\n\nEvery endpoint expects a Keycloak bearer token (realm `glmproject`, issuer `http://localhost:9080/realms/glmproject`).\n\nThe token **must carry a `tenant` claim** naming a declared company: each company owns a PostgreSQL schema, and the\nwhole `/api/**` surface answers `403` to a token without a known tenant. No endpoint takes the company as a\nparameter — it is always read from the token.\n\n## Roles\n\n- `USER` — shop-floor operator: clocks presence and work, reads.\n- `GESTIONNAIRE` — back-office: everything a `USER` can do, plus committing elements to the shop floor, closing\n  them, and correcting entries (`regularise`, `annule`, `corrige`).\n- `ADMIN` — technical administration (`/api/admin/**`, `/management/**`) only. **It grants no business access.**\n\nDevelopment users (password equal to the login): `gestionnaire.impeccmold`, `user.impeccmold`,\n`gestionnaire.katilys`, `user.katilys`.\n",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "glmproject API",
    "version" : "0.0.1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/elements-de-fabrication" : {
      "get" : {
        "operationId" : "list_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "debut",
          "required" : true,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "fin",
          "required" : true,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageRestElementDeFabrication"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "element-de-fabrication-resource" ]
      },
      "post" : {
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestCreationElementDeFabrication"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestElementDeFabrication"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "element-de-fabrication-resource" ]
      }
    },
    "/api/elements-de-fabrication/{id}" : {
      "delete" : {
        "operationId" : "delete_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "element-de-fabrication-resource" ]
      },
      "get" : {
        "operationId" : "get_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestElementDeFabrication"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "element-de-fabrication-resource" ]
      },
      "put" : {
        "operationId" : "update_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestModificationElementDeFabrication"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestElementDeFabrication"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "element-de-fabrication-resource" ]
      }
    },
...
    "/api/postes-de-travail" : {
      "get" : {
        "description" : "La page demandee, triee par libelle. Le filtre sur la nature est facultatif ; absent, il ne filtre rien.\n",
        "operationId" : "list",
        "parameters" : [ {
          "in" : "query",
          "name" : "nature",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageRestPosteDeTravail"
                }
              }
            },
            "description" : "La page demandee, triee par libelle."
          }
        },
        "summary" : "Lister les postes de travail",
        "tags" : [ "Postes de travail" ]
      },
      "post" : {
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestCreationPosteDeTravail"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestPosteDeTravail"
                }
              }
            },
            "description" : "Le poste est declare."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestPosteDeTravail"
                }
              }
            },
            "description" : "Un autre poste porte deja ce libelle."
          }
        },
        "summary" : "Declarer un poste de travail",
        "tags" : [ "Postes de travail" ]
      }
    },
    "/api/postes-de-travail/{id}" : {
      "delete" : {
        "description" : "Refuse tant qu'un operateur y est habilite : le supprimer le laisserait pointer sur du vide. Le retirer de leurs\nhabilitations d'abord.\n",
        "operationId" : "delete",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Le poste est supprime."
          },
          "404" : {
            "description" : "Poste de travail introuvable."
          },
          "409" : {
            "description" : "Des operateurs sont encore habilites sur ce poste, ou du temps y a deja ete pointe."
          }
        },
        "summary" : "Supprimer un poste de travail",
        "tags" : [ "Postes de travail" ]
      },
      "get" : {
        "operationId" : "get",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestPosteDeTravail"
                }
              }
            },
            "description" : "Poste de travail introuvable."
          }
        },
        "summary" : "Consulter un poste de travail",
        "tags" : [ "Postes de travail" ]
      },
      "put" : {
        "description" : "Le poste renomme s'affiche renomme partout : rien n'en est copie sur les operateurs, qui n'en retiennent que\nl'identifiant.\n",
        "operationId" : "update",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestModificationPosteDeTravail"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestPosteDeTravail"
                }
              }
            },
            "description" : "Poste de travail introuvable."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestPosteDeTravail"
                }
              }
            },
            "description" : "Un autre poste porte deja ce libelle."
          }
        },
        "summary" : "Reviser un poste de travail",
        "tags" : [ "Postes de travail" ]
      }
    }
  },
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
...
}